# manoward/static/sw/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Manowar 2.x API",
    "description": "Manowar 2 API. Designed to be a progrmattic way to interact with the data\ncollected and analyzed by Jellyfish 2.\n",
    "version": "2.6"
  },
  "servers": [
    {
      "url": "/manowar/v2"
    }
  ],
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/ip/search/": {
      "x-cached-length": "Every Midnight",
      "get": {
        "description": "Get the host(s) associated with this particular IP address\n",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "IP"
        ],
        "parameters": [
          {
            "name": "ip",
            "in": "query",
            "description": "The ipv4 or ipv6 address that you wish to search for records against.\n",
            "schema": {
              "type": "string",
              "format": "ipv46"
            }
          },
          {
            "name": "subnet",
            "in": "query",
            "description": "The ipv4 or ipv6 range of addresses that you wish to search against.\nthis can be combined with ip but it's not reccomended as only the ip\nwill match and only if it's in this range.\n",
            "schema": {
              "type": "string",
              "format": "subnet46"
            }
          },
          {
            "name": "iptype",
            "in": "query",
            "description": "Look for only one of the ips of this type. Must be vips4, vips6, host4,\nhost6, drac4, drac6, netdev4, netdev6 or unknown.\n",
            "schema": {
              "type": "string",
              "enum": [
                "vips4",
                "vips6",
                "host4",
                "host6",
                "drac4",
                "drac6",
                "netdev4",
                "netdev6",
                "unknown"
              ]
            },
            "required": false
          },
          {
            "name": "hostid",
            "in": "query",
            "description": "hostid to search for\n",
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "hostname",
            "in": "query",
            "description": "A regex to match for the hostname. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the hostname column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "A regex to match for the status. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the status column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pop",
            "in": "query",
            "description": "A regex to match for the pop. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the status column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "srvtype",
            "in": "query",
            "description": "A regex to match for the srvtype. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the status column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource",
            "in": "query",
            "description": "Matches the MOWN Resource either via Regex or Exactly (with exact). Keep in mind resources include\nthe resource qualifier and resource subtype if applicable.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "partition",
            "in": "query",
            "description": "Matches the MOWN Parition either via Regex or Exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service",
            "in": "query",
            "description": "Matches the MOWN Service either via Regex or Exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountid",
            "in": "query",
            "description": "Matches the MOWN Account ID either via Regex or Exactly\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mownbase",
            "in": "query",
            "description": "Matches the MOWN Base URI fully (with regex or Exactly). Does not include\na matching of arguments\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mownfull",
            "in": "query",
            "description": "Matches the MOWN full URI fully (with regex or Exactly). Includes a Matching of Arguments\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagged",
            "in": "query",
            "description": "Matches an individual tag. Returns true if the tag is defined (even if the tag is set to something\nlike false). Specify using JQ style formatting. Uses the [JSON_EXISTS](https://mariadb.com/kb/en/json_exists/)\nfunctionality to do the JSON mapping. So **if you're looking for tag `loveit` you'd want to speicfy `$.loveit`**.\n",
            "required": false,
            "schema": {
              "type": "string"
            },
            "format": "jsonpath",
            "pattern": "^\\$\\."
          },
          {
            "name": "taggedtrue",
            "in": "query",
            "description": "Matches an individual tag. Returns true if the tag is defined as true Uses the \n[JSON_EXTRACT](https://mariadb.com/kb/en/json_extract/)\nfunctionality to do the JSON mapping. So **if you're looking for tag `loveit` you'd want to speicfy `$.loveit`**.\nWill match against [true](https://mariadb.com/kb/en/true-false/), a builtin boolean. So if you specify something\nlike `{\"loveit\" : \"false\"}` (Note the String not boolean) it will return true and if you have something like\n`{\"zerosareawesome\" : 0}` it should return false.\n",
            "required": false,
            "schema": {
              "type": "string"
            },
            "format": "jsonpath",
            "pattern": "^\\$\\.\""
          },
          {
            "name": "exact",
            "in": "query",
            "description": "Controls your other filters and flips them from regex matches (default) to exact matches\n",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                true,
                false
              ]
            }
          }
        ]
      }
    }
  }
}